==> notebooks/1.0-pvso-initial-data-exploration.ipynb <==
# Importação

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns

# lendo a base de dados

df = pd.read_csv('../data/external/salaries.csv')

# Mostrando alguns registros da base de dados

df.head()

# Mostrando os tipos dos dados da base de dados

df.info()

# Resumo estatístico dos dados númericos da base de dados

df.describe()

# Mostrando o tamanho da base de dados (linhas e colunas)

df.shape

# Mostrando de forma pythonica o tamanho da base de dados
l, c = df.shape[0], df.shape[1]

print('Tamanho da base de dados')
print(f'Qtd linhas: {l}')
print(f'Qtd colunas: {c}')

# Mostrando as colunas da base de dados

df.columns

# Verificando o tipo de dado de cada coluna

df.dtypes

# Verificando se há registros nulos

df.isnull().sum()

# Resumo estatístico da base de dados com dados categoricos
df.describe(include='object')

# top exibe as informações mais frequentes em relação a coluna
# freq exibe a quantidade que as informações mais frequentes está repetindo
