==> notebooks/4.0-pvso-entrega.py <==
# Importação

import os
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import plotly.express as px
import pycountry
import streamlit as st

# --- Configuração da Página ---
# Define o título da página, o ícone e o layout para ocupar a largura inteira.
st.set_page_config(
    page_title="Dashboard de Salários na Área de Dados",
    page_icon="📊",
    layout="wide",
)

# --- Carregamento dos dados ---
caminho_base = os.path.abspath(os.path.join(os.path.dirname(__file__), '..'))
caminho_arquivo = os.path.join(caminho_base, 'data', 'processed', 'salarios_final.csv')
df = pd.read_csv(caminho_arquivo)
# df = pd.read_csv('../data/processed/salarios.csv')

# --- Barra Lateral (Filtros) ---
st.sidebar.header("🔍 Filtros")

# Filtro de Ano
anos_disponiveis = sorted(df['ano'].unique())
anos_selecionados = st.sidebar.multiselect("Ano", anos_disponiveis, default=anos_disponiveis)

# Filtro de Senioridade
senioridades_disponiveis = sorted(df['nivel_experiencia'].unique())
senioridades_selecionadas = st.sidebar.multiselect("Senioridade", senioridades_disponiveis, default=senioridades_disponiveis)

# Filtro por Tipo de Contrato
contratos_disponiveis = sorted(df['contrato'].unique())
contratos_selecionados = st.sidebar.multiselect("Tipo de Contrato", contratos_disponiveis, default=contratos_disponiveis)

# Filtro por Tamanho da Empresa
tamanhos_disponiveis = sorted(df['tamanho_empresa'].unique())
tamanhos_selecionados = st.sidebar.multiselect("Tamanho da Empresa", tamanhos_disponiveis, default=tamanhos_disponiveis)

# --- Filtragem do DataFrame ---
# O dataframe principal é filtrado com base nas seleções feitas na barra lateral.
df_filtrado = df[
    (df['ano'].isin(anos_selecionados)) &
    (df['nivel_experiencia'].isin(senioridades_selecionadas)) &
    (df['contrato'].isin(contratos_selecionados)) &
    (df['tamanho_empresa'].isin(tamanhos_selecionados))
]

# --- Conteúdo Principal ---
st.title("🎲 Dashboard de Análise de Salários na Área de Dados")
st.markdown("Explore os dados salariais na área de dados nos últimos anos. Utilize os filtros à esquerda para refinar sua análise.")
# st.markdown("---")

# --- Métricas Principais (KPIs) ---
st.subheader("Métricas gerais (Salário anual em USD)")

if not df_filtrado.empty:
    salario_medio = df_filtrado['salario_dolar'].mean()
    salario_maximo = df_filtrado['salario_dolar'].max()
    total_registros = df_filtrado.shape[0]
    cargo_mais_frequente = df_filtrado["cargo"].mode()[0]
else:
    salario_medio, salario_mediano, salario_maximo, total_registros, cargo_mais_comum = 0, 0, 0, ""

col1, col2, col3, col4 = st.columns(4)
col1.metric("Salário médio", f"${salario_medio:,.0f}")
col2.metric("Salário máximo", f"${salario_maximo:,.0f}")
col3.metric("Total de registros", f"{total_registros:,}")
col4.metric("Cargo mais frequente", cargo_mais_frequente)

st.markdown("---")

# --- Análises Visuais com Plotly ---
st.subheader("Gráficos")

col_graf1, col_graf2 = st.columns(2)

with col_graf1:
    if not df_filtrado.empty:
        top_cargos = df_filtrado.groupby('cargo')['salario_dolar'].mean().nlargest(10).sort_values(ascending=True).reset_index()
        grafico_cargos = px.bar(
            top_cargos,
            x='salario_dolar',
            y='cargo',
            orientation='h',
            title="Top 10 cargos por salário médio",
            labels={'salario_dolar': 'Média salarial anual (USD)', 'cargo': ''}
        )
        grafico_cargos.update_layout(title_x=0.1, yaxis={'categoryorder':'total ascending'})
        st.plotly_chart(grafico_cargos, use_container_width=True)
    else:
        st.warning("Nenhum dado para exibir no gráfico de cargos.")

with col_graf2:
    if not df_filtrado.empty:
        grafico_hist = px.histogram(
            df_filtrado,
            x='salario_dolar',
            nbins=30,
            title="Distribuição de salários anuais",
            labels={'salario_dolar': 'Faixa salarial (USD)', 'count': ''}
        )
        grafico_hist.update_layout(title_x=0.1)
        st.plotly_chart(grafico_hist, use_container_width=True)
    else:
        st.warning("Nenhum dado para exibir no gráfico de distribuição.")

col_graf3, col_graf4 = st.columns(2)

with col_graf3:
    if not df_filtrado.empty:
        remoto_contagem = df_filtrado['regime_trabalho'].value_counts().reset_index()
        remoto_contagem.columns = ['tipo_trabalho', 'quantidade']
        grafico_remoto = px.pie(
            remoto_contagem,
            names='tipo_trabalho',
            values='quantidade',
            title='Proporção dos tipos de trabalho',
            hole=0.5
        )
        grafico_remoto.update_traces(textinfo='percent+label')
        grafico_remoto.update_layout(title_x=0.1)
        st.plotly_chart(grafico_remoto, use_container_width=True)
    else:
        st.warning("Nenhum dado para exibir no gráfico dos tipos de trabalho.")

with col_graf4:
    if not df_filtrado.empty:
        df_ds = df_filtrado[df_filtrado['cargo'] == 'Data Scientist']
        media_ds_pais = df_ds.groupby('residencia_iso3')['salario_dolar'].mean().reset_index()
        grafico_paises = px.choropleth(media_ds_pais,
            locations='residencia_iso3',
            color='salario_dolar',
            color_continuous_scale='rdylgn',
            title='Salário médio de Cientista de Dados por país',
            labels={
                'salario_dolar': 'Salário médio (USD)',
                'residencia_iso3': 'País'
            }
                                       )
        grafico_paises.update_layout(title_x=0.1)
        st.plotly_chart(grafico_paises, use_container_width=True)
    else:
        st.warning("Nenhum dado para exibir no gráfico de países.")

with st.container():
    if not df_filtrado.empty:
        df_ds = df_filtrado[df_filtrado['cargo'] == 'Data Scientist']
        salario_medio_experiencia_regime = df_ds.groupby(
            ['ano','nivel_experiencia', 'regime_trabalho']
            )['salario_dolar'].mean().reset_index()
        
        fig = px.bar(
            salario_medio_experiencia_regime,
            x='ano',
            y='salario_dolar',
            color='nivel_experiencia',
            labels={
                'salario_medio': 'Salário Médio',
                'ano': 'Ano',
                'nivel_experiencia': 'Nível de Experiência',
                'regime_trabalho': 'Regime de Trabalho'
            },
            facet_col='regime_trabalho',
            barmode='group',
            title='Salário Médio(USD) por Nível de Experiência e Regime de Trabalho ao Longo dos Anos'
        )

        fig.update_layout(
            yaxis_title='Salário Médio',
            xaxis_title='Ano'
        )

        # fig.show()
        # fig.update_layout(title_x=0.1)
        st.plotly_chart(fig, use_container_width=True,key='Sálario médio(USD) por nível de experiência por Regime')
    else:
        st.warning("Nenhum dado para exibir no gráfico de países.")

with st.container():
    if not df_filtrado.empty:
        # df_ds = df_filtrado[df_filtrado['cargo'] == 'Data Scientist']
        salario_medio_cargo_empresa = df.groupby(
            ['cargo', 'tamanho_empresa']
        )['salario_dolar'].mean().reset_index()

        # salario_medio_cargo_empresa

        top10_cargos = (
            salario_medio_cargo_empresa.sort_values(['tamanho_empresa', 'salario_dolar'], ascending=[True, False])
            .groupby('tamanho_empresa').head(10)
        )
        
        fig = px.bar(
            top10_cargos,
            x='tamanho_empresa',
            y='salario_dolar',
            color='cargo',  # barras lado a lado por tamanho de empresa
            barmode='group',
            # orientation='h',
            labels={'salario_dolar': 'Salário Médio', 'cargo': 'Cargo', 'tamanho_empresa': 'Tamanho da Empresa'},
            title='Top 10 Cargos com Maior Salário Médio por Tamanho de Empresa'
        )

        fig.update_layout(
            yaxis_title='Salário Médio',
            xaxis_title='Ano'
        )

        # fig.show()
        # fig.update_layout(title_x=0.1)
        st.plotly_chart(fig, use_container_width=True, key='Top10 cargos por tamanho de empresa')
    else:
        st.warning("Nenhum dado para exibir no gráfico de países.")

st.markdown("---")

# --- Tabela de Dados Detalhados ---
st.subheader("Dados Detalhados")
st.dataframe(df_filtrado)
